requirements.txt
langchain_groq==0.1.9
python-dotenv
crewai
crewai_tools